requirements.txt
apify-client
beautifulsoup4
flask
google-auth
google-generativeai
google-genai
gspread
httpx
requests
